# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Siriusx-API
  description: |
    Siriusx-API 是一个智能 API 网关，用于聚合多个 Claude 供应商，提供统一的 OpenAI 兼容接口。

    ## 主要功能
    - 🔄 格式转换：Claude ↔ OpenAI 格式互转
    - 🏢 供应商管理：多供应商配置和负载均衡
    - 🎯 模型映射：统一模型到多供应商映射
    - 🔐 Token 管理：API 令牌生成和验证
    - 📊 监控统计：系统状态和请求统计

  version: 1.0.0
  contact:
    name: Siriusx-API Team
    url: https://github.com/Mieluoxxx/Siriusx-API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: 本地开发环境
  - url: https://api.siriusx.com
    description: 生产环境
...
paths:
...
  /api/providers:
    get:
      summary: 获取供应商列表
      description: 获取所有供应商的列表
      tags:
        - Providers
      responses:
        '200':
          description: 成功返回供应商列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Provider'

    post:
      summary: 创建供应商
      description: 添加新的 Claude API 供应商
      tags:
        - Providers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProviderRequest'
      responses:
        '201':
          description: 成功创建供应商
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Provider'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/providers/{id}:
    parameters:
      - $ref: '#/components/parameters/ProviderId'

    get:
      summary: 获取单个供应商
      description: 根据 ID 获取供应商详细信息
      tags:
        - Providers
      responses:
        '200':
          description: 成功返回供应商信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Provider'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      summary: 更新供应商
      description: 更新供应商的配置信息
      tags:
        - Providers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProviderRequest'
      responses:
        '200':
          description: 成功更新供应商
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Provider'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: 删除供应商
      description: 删除指定的供应商
      tags:
        - Providers
      responses:
        '204':
          description: 成功删除供应商
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/providers/{id}/enabled:
    parameters:
      - $ref: '#/components/parameters/ProviderId'

    patch:
      summary: 启用/禁用供应商
      description: 切换供应商的启用状态
      tags:
        - Providers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                enabled:
                  type: boolean
                  description: true 为启用，false 为禁用
                  example: true
      responses:
        '200':
          description: 成功更新状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Provider'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    ProviderId:
      name: id
      in: path
      required: true
      description: 供应商 ID
      schema:
        type: integer
        example: 1
...
  schemas:
...
    Provider:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "OneAPI Provider"
        base_url:
          type: string
          format: uri
          example: "https://api.oneapi.com"
        api_key:
          type: string
          description: API Key (脱敏显示)
          example: "sk-****5678"
        priority:
          type: integer
          description: 优先级（数字越大优先级越高）
          example: 80
        enabled:
          type: boolean
          description: 是否启用
          example: true
        healthy:
          type: boolean
          description: 健康状态
          example: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateProviderRequest:
      type: object
      required:
        - name
        - base_url
        - api_key
      properties:
        name:
          type: string
          maxLength: 100
          example: "OneAPI Provider"
        base_url:
          type: string
          format: uri
          example: "https://api.oneapi.com"
        api_key:
          type: string
          example: "sk-abc123def456"
        priority:
          type: integer
          minimum: 0
          maximum: 100
          default: 50
          example: 80
    UpdateProviderRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        base_url:
          type: string
          format: uri
        api_key:
          type: string
        priority:
          type: integer
          minimum: 0
          maximum: 100
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "Invalid request parameters"
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "NOT_FOUND"
              message: "Resource not found"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        使用 API Token 进行认证。Token 格式为 `sk-` 开头的 Base64 编码字符串。

        示例：`Authorization: Bearer sk-soj6C7X3f3uZCkUSN27I9qSb6UjM61nxXeHeBKAquMU=`
security:
  - BearerAuth: []